+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 18. Added Start Screen Text Color Customization.
 19. Fixed Issue with Colors Reading and Writing to Start Screen Config Incorrectly.
 20. Implemented Element Toggles for Start Screen.
+21. Added LemonArraySegment because System.ArraySegment doesn't have interfaces on older .NET Framework versions.
+22. Fixed Issue with SharpZipLib's InflaterHuffmanTree not properly Segmenting Array.
 
 ---